requirements.txt

fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
scikit-learn
numpy
prometheus-client
aiohttp
openai
tqdm
python-pptx
